Remove mdRef@XPTR attribute
Not used in any registered profiles; mdRef@href is sufficient to link to
metadata files. There are no live examples of XPTR usage in METS in a
quick Google search, just documentation and libraries that try to
completely cover the METS standard.

N.B. It is in use in HathiTrust METS, but it is not used correctly there
and should not be difficult to migrate to something more standard.