Add mdGrp for grouping related metadata
This makes mdSec / mdGrp / md parallel to fileSec / fileGrp / file

- make mdSec non-repeatadble
- Rename metadata element to md; metadataType to mdType
- change references to amdSec to mdSec
- remove lingering DMDID attribute

For now, mdGrp allows only the ID attribute, but we will almost
certainly want to add others. In particular we may want to add USE and
MDID (e.g.  for referencing provenance metadata about how or when the
metadata in this mdGrp was created), and we should discuss further how
to handle USE for the md element -- currently it represents the same
thing as the element name for things of mdSecType in the old schema
(i.e.  techMD/rightsMD/sourceMD/digiprovMD/dmdSec), but it seems like we
should probably either remove the restriction and have an external
suggested list of values, or move it to another attribute.